feat(memory): 添加表前缀配置并调整相似度阈值类型- 在 MemoryConfig 中添加 TablePre 字段，用于配置表前缀 - 将 Threshold 类型从 float32 改为 float64，提高精度 - 在 MemoryStorage 接口中添加 SetTablePrefix 方法 - 实现 MemoryStore 和 SQLStore 的 SetTablePrefix 方法- 修改 SQL 模型的 TableName 方法，使用表前缀 - 新增 TableNameProvider 结构体和相关方法，用于生成带前缀的表名

+++ agent/agent.go
@@ -52,7 +52,7 @@ type KnowledgeQueryConfig struct {
 	// 最大查询结果数量
 	MaxResults int
 	// 相似度阈值
-	Threshold float32
+	Threshold float64
 	// 是否总是查询（不使用关键词触发）
 	AlwaysQuery bool
 }
